pyproject.toml
[project]
name = "geminiapi-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.7.0",
    "rich>=13.9.4",
    "streamlit>=1.43.2",
]
